requirements.txt
antlr4-python3-runtime==4.8
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
atari-py==0.2.6
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
Box2D==2.3.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.2
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
gym==0.18.0
hydra-core==1.0.6
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
omegaconf==2.0.6
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.3
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pyglet==1.5.0
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
scipy==1.6.1
Send2Trash==1.5.0
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
terminado==0.9.2
testpath==0.4.4
torch==1.7.1
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.10.21
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
